+++ sshd-core/src/main/java/org/apache/sshd/common/forward/Forwarder.java
@@ -28,7 +28,7 @@ import org.apache.sshd.common.util.net.SshdSocketAddress;
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public interface ForwardingFilter
+public interface Forwarder
         extends PortForwardingManager,
         PortForwardingEventListenerManager,
         PortForwardingEventListenerManagerHolder,