+++ CHANGES.md
@@ -8,11 +8,11 @@
 
 # [Version 2.5.0 to 2.5.1](./docs/changes/2.5.1.md)
 
-# [Version 2.5.1 to 2.6.0](./docs/changes/2.6.0.md)
-
 # Planned for next version
 
 ## Major code re-factoring
+
+* [SSHD-1034](https://issues.apache.org/jira/browse/SSHD-1034) Rename {{org.apache.sshd.common.ForwardingFilter}} to {{Forwarder}}.
 
 ## Minor code helpers